chore(tooling): add typescript
