--- index.js
@@ -1,3 +1,0 @@
-import server from "./src/server";
-
-server();
