+++ package.json
@@ -1,11 +1,17 @@
 {
   "name": "mockserver",
   "version": "0.0.1-pre",
-  "description":
-    "A mockserver that allows you to mock XHR, long-polling XHR, server sent events and websocket connections",
-  "main": "index.js",
+  "description": "A mockserver that allows you to mock XHR, long-polling XHR, server sent events and websocket connections",
+  "main": "dist/index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "node dist/index.js",
+    "prestart": "npm run build",
+    "build": "npm run build-ts",
+    "prewatch": "npm run build",
+    "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"cyan.bold,green.bold\"  \"npm run watch-ts\" \"npm run watch-node\"",
+    "watch-node": "nodemon dist/index.js",
+    "build-ts": "tsc",
+    "watch-ts": "tsc -w",
     "setup:git-hooks": "./scripts/pre-install",
     "commitlint": "commitlint -e",
     "test": "jest",
@@ -22,28 +28,73 @@
   ],
   "author": "DanielMSchmidt <dschmidt@mesosphere.com>",
   "license": "MIT",
+  "homepage": "https://github.com/mesosphere/mockserver#readme",
   "bugs": {
     "url": "https://github.com/mesosphere/mockserver/issues"
   },
   "jest": {
     "projects": [
       {
-        "displayName": "Jest"
+        "displayName": "Jest",
+        "transform": {
+          "^.+\\.ts$": "ts-jest"
+        },
+        "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(js|ts)$",
+        "moduleFileExtensions": [
+          "ts",
+          "js",
+          "json"
+        ]
       },
-      { "displayName": "Prettier", "runner": "jest-runner-prettier", "moduleFileExtensions": ["js", "jsx", "json", "ts", "tsx", "css", "less", "scss", "graphql", "md", "markdown"], "testMatch": ["**/*.js", "**/*.jsx", "**/*.ts", "**/*.tsx", "**/*.css", "**/*.less", "**/*.scss", "**/*.graphql", "**/*.md", "**/*.markdown"] }
+      {
+        "displayName": "Prettier",
+        "runner": "jest-runner-prettier",
+        "moduleFileExtensions": [
+          "js",
+          "ts",
+          "md"
+        ],
+        "testMatch": [
+          "src/**/*.js",
+          "src/**/*.ts",
+          "**/*.md"
+        ]
+      },
+      {
+        "displayName": "TsLint",
+        "runner": "jest-runner-tslint",
+        "moduleFileExtensions": [
+          "ts"
+        ],
+        "testMatch": [
+          "**/*.ts"
+        ]
+      }
     ]
   },
-  "homepage": "https://github.com/mesosphere/mockserver#readme",
   "dependencies": {
     "express": "^4.16.3",
     "npmlog": "^4.1.2"
   },
   "devDependencies": {
+    "@dcos/tslint-config": "0.0.1",
+    "@types/express": "^4.11.1",
+    "@types/jest": "^22.2.3",
     "babel-jest": "^22.4.3",
     "babel-preset-env": "^1.6.1",
+    "concurrently": "^3.5.1",
     "detect-port": "^1.2.2",
     "jest": "^22.4.3",
     "jest-runner-prettier": "^0.2.2",
-    "prettier": "^1.12.0"
+    "jest-runner-tslint": "^1.0.4",
+    "nodemon": "^1.17.3",
+    "prettier": "^1.12.0",
+    "ts-jest": "^22.4.3",
+    "tslint": "^5.9.1",
+    "tslint-config-prettier": "^1.10.0",
+    "tslint-eslint-rules": "^5.1.0",
+    "tslint-plugin-prettier": "^1.3.0",
+    "tslint-react": "^3.5.1",
+    "typescript": "^2.8.1"
   }
 }
