+++ src/__tests__/config-test.ts
@@ -1,15 +1,15 @@
-import { getPort } from "../server";
+import { getPort } from "../config";
 
-describe("Server", () => {
+describe("Config", () => {
   describe("#getPort", () => {
     it("returns a default port if no port is configured", () => {
       expect(getPort()).toBe(3000);
     });
 
     it("returns the configured port", () => {
-      process.env.PORT = 4002;
+      process.env.PORT = "4002";
       expect(getPort()).toBe(4002);
-      process.env.PORT = null;
+      process.env.PORT = undefined;
     });
   });
 });
