+++ src/config.ts
@@ -1,0 +1,3 @@
+export function getPort(): number {
+  return process.env.PORT ? parseInt(process.env.PORT, 10) : 3000;
+}
