+++ src/index.ts
@@ -1,0 +1,3 @@
+import server from "./server";
+
+server();
